checking inference results for loose orientations and mean_flip
A new sequential analysis on source data reconstructed with loose=0.2,
pick_ori='normal' and label_mode='mean_flip' with inference options as in the
previous analysis is available in

	source_sequential_201807041930.h5

The results are very similar to previous results with fixed orientations. So
neither loosening the orientation constraints (slightly), nor flipping the
signs of source currents of opposing flanks of a gyrus before averaging within
areas has a big impact on the results. I expected more areas to exhibit
significant results with the flipping, but this is also not really the case.
Effects in pre-motor areas (6a?) became stronger and the effects in M1 became a
bit weaker, but overall the areas with significant effects after multiple
comparison correction were very similar.

Checking the source plot for individual areas (Figure 7 in the paper), I was
surprised that the time courses even of individual participants are almost
exactly the same as before. There are some differences, but these are hard to
spot. Perhaps this is only the case for the most strongly activated areas that
I looked at (which are also the same as before: R-v23ab and R-POS2).

Changes from previous results that I noticed: In the grand average effect
magnitude time course the location of the peaks changed slightly as recorded in
make_source_figures.py.

Accumulated evidence
--------------------
Checking the time courses of the remaining regressors I got the idea to isolate
areas involved in evidence accumulation by selecting those areas which show a
correlation with accumulated evidence at a time point in which those areas do
not show a correlation with dot_x. This occurs at 160 ms after dot onset
corresponding to 60 ms in the time frame of the (previous) accumulated evidence
regressor. The only area with a significant effect at that time point is left
area 2 in the somatosensory and motor cortex.