recompute response-aligned effects for M1 with new source reoncstruction
Given that the other results were quite similar as before I did not expect big
changes, but because the effects in M1 were a bit lower in the sequential
analysis, I feared that there would be fewer significant effects before the
response. This is approximately true, but the overall picture did not change.