new source-level inference in motor vertices inference
For completeness I did the source reconstruction with loose=0.2 and
pick_ori='normal' for the motor vertices. Results are in

	source_epochs_allsubs_HCPMMP1_5_8_201807051740.h5

In light of the similarity of results for the within-area averaged data, I do
not expect any big changes in correlation effects for the individual sources
either. Inference is currently running.