understanding MNE label sign flips with motor vertices results
MNE's sign flipping does not quite work as I expected. It computes for each
label separately something like the mean orientation of the vertex normals
(through a PCA-approach) and then flips the sign of the vertex value for those
vertices for which their normals are rotated by more than 90 degrees in either
direction from the mean orientation. This means that labels with similar
vertex normal orientations throughout the area tend to not have any sign flips.
For some reason I expected values lying on different sides of a gyrus being
appropriately flipped, but this is not true, if the two sides of the gyrus
belong to different labels/areas. The MNE approach automatically handles
complicated geometry of the cortex surface, but is really just a within-label
heuristic.

This means that even after applying label sign flipping, the sign of
correlations with dot_x in the motor vertices follows the anatomical shape of
the cortex, because the motor areas roughly follow the gyral landscape and
consequently have a similar orientation across their vertices. I probably could
produce a sign flipping the way I expected it by computing the sign flips
across all the areas of interest instead of only the individual areas. I
wonder, however, whether this is really what I want to eventually do, given
that the whole sign flipping business is heuristic. That the currents appear to
flow in a common direction irrespective of the orientation of the cortical
surface is an insight on its own which at least points to the insufficient
spatial resolution of MEG.