surprising flip patterns in motor vertices
Even though I think to have understood how label sign flips are generated now,
I'm surprised about the actual flip patterns generated by the process.
Following my intuition that the overall orientation of the cortex should vary
along the anterior-posterior gradient, I thought that giving all motor vertices
to MNE's label flip function would flip the signs of values based on this
gradient, i.e., the flip would equalise the signs of effects between different
sides of the precentral and postcentral gyri. This is not the case. Even when
restricting the combination of areas to the respective hemispheres, values in
primary motor cortex are flipped even though they lie on a common side of the
precentral gyrus. The pattern of flips rather appears to follow other
orientation gradients, but it's really hard to pin down what exactly happens,
even when plotting the white matter surface which shows the anatomical
structure more clearly.

I have at least verified that the mean orientation of the combined label on
which the sign flips are based is similar to the mean orientation of the
primary motor cortex alone, i.e., my initial intuition was not so bad. The
differences I observed are still apparently sufficient to introduce weird flip
patterns. Seeing that the flip pattern of M1 only did not mingle up correlation
signs within M1 much, I could simply use the M1 mean orientation as comparison
and apply it to the remainder of the areas.