comparing response-aligned effects across motor areas
In discussion with Stefan we agreed that the pattern of signs of correlations
across motor vertices could simply mean that the source reconstruction is
unfortunately not very spatially precise, as I had intuited previously. To
investigate how specific the correlations in primary motor cortex are to
primary motor cortex, I repeated the response aligned analysis in the window
[300-500] ms after dot onset for areas around the motor cortex which also had
correlations effects in this time window. Results can be found in:

	source_response-aligned_201807101536.h5 (area 3a)
	source_response-aligned_201807101714.h5 (area 6a)
	source_response-aligned_201807101839.h5 (area FEF)

I included FEF as a potentially interesting area that did not show pronounced
correlation effects in the dot-aligned analysis.

As expected, areas 3a and 6a show similar patterns of response-aligned effects
as the primary motor cortex (area 4). Specifically, they both exhibit
significant effects around the response, but also up to 250 ms before. Overall,
however, area 6a appears to have weaker effects and especially lacks
significant effects earlier than 250 ms before the response (except for 1 time
point in the left hemisphere and 3 in the right). Area 3a, on the other hand,
exhibits almost the same time course of effects as the primary motor cortex
(except that the sign is flipped, the right hemisphere is considerably weaker
and the very early effects before -700 ms are not significant for 3a). Perhaps
this is sufficient to argue that the effects are strongest in, or very close to
primary motor cortex.