functional connectivity through delayed correlations
I wanted to make better use of the high temporal resolution of MEG and
investigate the dynamics of information processing, or rather transmission
across brain areas. My idea was to correlate the raw source currents of one
brain area with those of another brain area at another time point where the
correlation is across trials. Further, I wanted to concentrate on the
dot-specific effects and decided to expand the correlation data sets with 100
ms spaced time points as I've done in the expanded regression analysis. This
approach should remove, or at least downweigh correlations only present at
specific time points.

I hoped that the strength of the correlations would peak at a specific latency
(delay) which would allow me to argue that information from one brain area is
transmitted to another in a specific order.

Results are still computing (~15 min per subject) and can be found in:
	source_crosscorr_201807131159.h5

I have only looked at subject 2 so far. The results are very hard to interpret.
The data expansion seems to work in the sense that I believe to recognise
dot-correlation patterns, e.g., for time points 120 and 180 ms. The downside of
this is that these patterns recur every 100 ms. Also, the correlation plots
show diagonal stripes going from the upper left to the lower right. These
indicate that there are time points in the comparison area which exhibit
specifically strong or weak correlations with the currents in the seed area at
a given time point. This is because, the plot shows an absolute time on the
x-axis (the seed time point) against a relative time on the y-axis (the delay
in the comparison area leading to an absolute time of seed time + delay).

The strongest correlations always occur closely around a delay of 0, even when
considering areas that are anatomically quite far apart such as V3 and 3a. This
could mean that the reconstructed source currents are strongly influenced by
global activation patterns on the sensor level that are spread across many
brain areas during reconstruction.

As expected, areas very close together such as 31pd and v23ab have very strong
correlations. They also don't exhibit much structure in the correlation plots
except for 100ms-spaced delay-dependent stripes reflecting the dot onset times.
These stripes are also present in correlations within the same area.

Some area combinations show more interesting correlation patterns. For example,
v23ab -> 6a. This has strong diagonal patterns with clearly strongest
correlations when considering transmission from left v23ab to right 6a, but I
do not have an interesting interpretation of these results that goes beyond
noticing that there is one time point at which the right 6a appears to be
specifically strongly correlated with v23ab.

The correlation patterns are always strongly repetitive across seed times and
delays, but become weaker for increasing delays. The latter might be simply due
to fewer data points contributing to the correlation coefficients. The only
difference between the correlations at seedt=200 and seedt=300 is that the 480
data points corresponding to t=200 are not included anymore, but all others
are. It may therefore make sense to only look at a 100 ms square of
correlations from these results.

Perhaps it's even better, i.e., easier to interpret, to not pool data across
time points.

I do also wonder why I actually see correlation patterns linked to 100 ms
spaced dot onsets. These should not reflect information about the x- or
y-coordinates of the dots, because the coordinates were (supposed to be)
uncorrelated across trials. So these patterns should just reflect dot-onset
related effects. If that is so, why do I actually see (smaller sized)
correlations between posterior cingulate and motor areas, or even early visual
and motor areas (e.g. V3 -> 3a)?

In summary, I have my doubts now about the usefulness of this analysis. It
appears that the correlations across areas just follow somehow a common average
signal that is generated at each dot onset. It is unclear to me, however, why
this signal should have to do anything with motor areas. If motor areas always
activate with dot onsets, then the signal should not be anti-correlated across
hemispheres, as it is.