investigating mean area cross-correlations
It turns out that the correlation patterns described in the last commit for
subject 2 can be quite different across subjects. I have now computed the mean
correlation across subjects and looked at that. Unfortunately my computations
of the correlations across areas halted at subject 29, because pandas could not
open the HDF5-file for appending new results anymore *). Therefore, everything
described below is from the average across subjects until subject 28.

Difference to subject 2 results
-------------------------------
From a few checks the mean correlation plates (correlations for one area pair
across seed times and delays for the 4 combinations of hemispheres) now have
generally lower correlation magnitudes. The structure in some plates also has
faded somewhat and the correlation for delay=0 remains as the one prominent
effect across all seed times.

Summary measures
----------------
I followed up my initial idea and identified the delay that has the largest
correlation magnitude for each combination of areas (maxcorr). For about 87% of
all combinations at seed time = 180 this is delay=0. Those area combinations
that have the peak correlation for a different delay tend to have much smaller
correlation magnitudes on average across delays (avmagcorr; the average
correlation magnitude tends to be one order of magnitude smaller). This
suggests that these results may be driven by noise.

I also computed the average correlation magnitude across seed times (at
delay=0) instead of delays (avmagcorrdelay), but the pattern of results is very
similar to that in avmagcorr. The reason for this is that the correlation
magnitudes at delay=0 do not actually differ much across seed times so that
taking the average across seed times just gives you approximately the magnitude
at the chosen delay and as the magnitudes are strongest at delay=0, the average
across delay will also be dominated by that delay.

Reconstructing a sensible arrangement of areas from correlations
----------------------------------------------------------------
It is clear from the results that areas lying close together in a brain region
also exhibit strongly correlated source currents across trials. I wondered
whether I could identify clusters of areas that functionally belong together.
To investigate this I used multidimensional scaling (MDS) to transform the
cross-correlation matrix into a spatial arrangement of areas. My hope was that
in this lower dimensional map functionally related brain areas would cluster
together.

It turned out that I cannot identify discrete clusters on that map. Instead the
areas roughly construct a sphere (in 3D). I haven't checked which brain areas
are neighbours in this map, but they at least roughly separate into left and
right hemisphere areas, as one would expect.

Conclusion
----------
Seeing that my across-trial cross-correlations do not vary much across seed
times strongly suggests that these correlations do not actually measure what I
intended to measure. Specifically, they appear to be strongly influenced by
across-trial variability that may be unrelated to the processing of the dots.
On the other hand, if that was true, shouldn't I have also observed that the
cross-correlations are roughly constant across delays? Yet, it is very clear in
most area combinations that even increasing or decreasing the delay by 30 ms
from 0 considerably reduces the magnitude of cross-correlations. This
combination of results suggests that the strength of the relatedness of the
signals in two areas does only depend on their relative timing, but not on the
specific time from dot onset. This excludes the possibility that the measured
cross-correlations are only driven by across trial variability of the signals
that does not change across time points within a trial (this should have been
very unlikely anyway as the signals were bandpass filtered and baseline
corrected), but it also means that these correlations have most likely nothing
to do with the processing of the dots, i.e., decision evidence. So it's still
unclear to me what these correlations represent.

One issue could be that the correlations that I have computed may also depend
on the within-trial dynamics of the signals, i.e., the correlations may be
influenced by how the signals change on average from one 100 ms step to the
next and whether this is the same in both considered areas. If the signal in
both areas equally rises linearly within trial, the resulting correlation may
indeed be independent of the seed time, but it would also be independent of the
delay. So the dynamics should be nonlinear to account for the observed pattern
of results. One explanation for that pattern could be that the dynamics is
oscillatory with a period of 100 ms (as dots were 100 ms apart). Then also the
particular time within the trial should be irrelevant for the magnitude of the
correlation. The delay, however, could determine the correlation. I have
verified this in a simulation, but for a perfect sine wave the correlation
should also reverse its direction as you change the delay to half the period of
the oscillation (leading to perfect anti-correlation instead of correlation). I
did not observe this in my results.

I'm inclined to try out the cross-correlation analysis without any within-trial
dynamic effects by properly re-epoching the data to dot onsets and applying
baseline correction to each newly generated epoch. This should remove parts of
the correlations contributed by, for example, rising signals. I'm not sure
whether this then allows me to investigate the desired transfer of information
from one area to another, because the strong preference for delay=0 still
suggests that most of the signals are strongly locked to the 100 ms timing of
the dots.

*) file: source_crosscorr_201807131159.h5
   probably the problem resulted from me reading and writing from the file at
   the same time