select desired labels early + dot-onset-specific baselines
The idea was to re-epoch the data onto dot onsets and subtract a
dot-onset-specific baseline from each of those to reduce the effects from other
dots on the signal for a chosen dot. I only ended up computing the baselines,
but did not re-epoch as it turned out that I would have to do this by hand and
I was not sure about the eventual benefits.