regression in sensor space with sum_dot_x instead of accev and dot_x
This is a check I wanted to make to see whether shifted accev and sum_dot_x
indeed give similar results, as expected given that dot_x and accev are almost
uncorrelated. Further, I was mostly interested to see how accumulated evidence
is represented before 100 ms. The results indicate that the accumulated
evidence can be explained by superposition of dot_x effects.

Results
-------
are in
	meg_sequential_201808091710.h5

The time courses of the two grand averages indeed are very clost to each other.
accev tends to be slightly smaller than sum_dot_x up to about 320 ms from where
this is reversed. The difference between sum_dot_x and accev is that accev is
offset by the model-estimated choice bias and that accev uses variation in
presented dots to explain data that was recorded 100 ms later in comparison to
sum_dot_x, e.g., the accumulated evidence after inclusion of dot 5 is
associated with time points 400 to 500 ms in sum_dot_x and time points 500 to
600 ms in accev (plus potential delay). I don't currently see how these
differences explain the pattern of higher and lower estimated correlation
magnitudes. Perhaps it has also something to do with not having included dot_x
into the analysis with sum_dot_x such that the sum_dot_x regressor picks up
effects that were explained by dot_x at the early times up to 320 ms, but this
doesn't explain why accev effects then exceed sum_dot_x effects from 320 ms.
Anyway, in relation to the overall effect magnitudes these differences are
smallish so that I intend to ignore them.

The grand average correlation magnitudes of sum_dot_x before 100 ms show
another peak at 80 ms with an additional trough at 20 ms. That's a bit
surprising, but there were hints in the single sensor plot. The peak at 80 ms
could reflect a correlation with the accumulated evidence of the previous dot
180 ms after occurrence of the previous dot. The reason why there is a trough
at 120 ms instead of a maintained correlation as indicated by the continuing
effects from 180 ms onwards is probably the anti-correlation of the signal with
dot_x at 120 ms. As accumulated evidence and dot_x are themselves correlated
this anti-correlation between dot_x and the signal eats up some of the
correlation of the signal with accumulated evidence. Similarly, this would
explain the trough at 20 ms.

The correlation between successive values of accumulated evidence do actually
explain the whole trajectory of the accumulated evidence effects. What we see
is essentially a superposition of dot_x effects for the previous and the
current dot (minus the anti-correlations at 120 ms). Under these circumstances
I wonder whether we can find representations of specifically accumulated
evidence at all. What would it take?