showing accumulated evidence correlations in relation to x-correlations
The idea was that the sum_dot_x effects could be described as a linear
combination of temporally shifted dot_x effects. I implemented this using a
GLM. This works well for the plateau of effects, but has issues in the first
200 ms, because there we have the subtraction of effects for 120 ms dot_x
effects so that a simple weighted sum cannot describe the reduction of
correlation. Additionally, the contribution of dot_x effects from different
dots could be different at different time points, e.g., at 500 ms the
contribution from the pre-previous dot x-coordinates could be less than at 120
ms which corresponds to the start of the dot_x plateau at 320 ms. So what I
plot in this committed version is simply an overlay of temporally shifted dot_x
effects on top of the sum_dot_x correlations without the weighting estimated by
the GLM.

This display shows that all the peaks in the sum_dot_x correlations can be
reasonably well explained by dot_x effects, as I have already hinted at in the
last commit. This discredits the accumulated evidence regressor as carrying
genuine addtional information, or at least it's hard to argue that this is the
case. We should not be surprised, however, after all sum_dot_x is just a sum of
dot_x. I wondered whether it would be possible to select trials so that
sum_dot_x and dot_x would be almost uncorrelated. This would be a kind of
post-hoc experimental design. Whatever I do, however, I might be able to reduce
the correlation with dot_x, but it will be impossible to simultaneously
minimise the correlations with previous dot_xs. Also, I wonder whether
sufficiently many data points would be left for identifying the small
correlations somewhat reliably.

Another alternative for identifying accumulated instead of momentary evidence
is that supposedly accumulated evidence reaches a fixed bound at decision time.
It could, however, be that some areas in the brain keep accumulating evidence
as long as evidence is coming in. The greater issue in my eyes is that the
accumulated evidence appears to be represented primarily in motor areas which
also signal the motor response after the decision. The recorded MEG signals are
then swamped by the motor signal making it hard or impossible to observe the
relatively tiny correlations with accumulated evidence and whether they reach a
common bound.