new figure comparing topographies between momentary and accumulated evidence
This is supposed to show that accumulated evidence (sum_dot_x) is located more
centrally. I did not expect that accumulated evidence has also stronger
correlations posterior laterally next to the parietal sensors at which
momentary evidence is represented more strongly, but in hindsight this can also
be seen in the original topographies.