new inference run with sum_dot_x in source space
This is for investigating the brain areas associated with accumulated evidence.
Results are in:
	source_sequential_201808131739.h5

Compared to the effects with dot_x there are fewer significant effects for
sum_dot_x. This appears to be due to an increased variability of parameter
estimates across subjects, because the average parameter estimates (betas) are
higher for sum_dot_x than for dot_x.