generalising average brain figures for shoing dot_x results across time windows
This is for the new main source reconstruction results figure in the paper. The
script now produces automatically all individual figures for the chosen time
windows for all hemispheres, but the produced figures still need to be stitched
together manually e.g. in Gimp afterwards to make them nice and add
annotations.

The figure generation process is now centred on the user choice of the
regressor of interest.

I chose FDR alpha = 0.05, because for sum_dot_x there are no significant
effects after correction with alpha = 0.01. For further discussion of results
see BeeMEG_v2.0.docx.