bugfix: use reindex instead of loc + no subject in rtindex for response-aligned
Both fixes were necessary to run response-aligned analyses. Because for
response-aligned times not all desired time points are present in the epoch
data store, I had to change to reindex from loc (reindex fills non-existent
data with nans). The code returning the response-aligned index also returned
the subject id inside the index. That was not compatible with the other code in
infer_singledot.py