control analysis: only using sign of dot_x and accev as regressors
I wondered whether we could say for the paper that the high resolution of
decision evidence helped us in identifying the representations of decision
evidence in the brain. So I introduced new regressors which only encode the
sign of dot_x and accev instead of the actual size of the x-locations. The
subsequent regression analysis, thus, only looks for signals that indicate the
side of the screen that the dot has been (for dot_x).

Results are in:
	source_sequential_201808161156.h5

I regenerated the brain figures that I recently produced for the paper with
dot_x_sign instead of dot_x and found that the identified brain areas are very
similar to the original analysis. The largest differences I see are that
sometimes different areas are among the top 5, but mostly the effects are
comparable with only rarely an additional brain area being identified with one
or the other regressor.

Interestingly, with the full dot_x we see significant effects in the early
window in somatosensory areas, but these are absent for dot_x_sign. Also, the
t-value magnitudes tend to be slightly higher for dot_x than dot_x_sign, at
least among the significant effects in the selected time windows (as indicated
by the colorbars). Does that mean that the full dot_x is more sensitive to
identify representations of decision evidence than dot_x_sign? This is a hard
sell. The t-value differences are really small (around 0.1 or lower). And why
should there be a representation of the x-coordinate / momentary evidence
already 120 ms after dot onset in somatosensory areas?

In total, this new analysis shows that the brain effects are strongly driven by
simply whether the dot was on the left or right of the screen. How far
individual dots were on either side appears to have mattered less.