response-aligned single-dot analysis in brain areas for choice correlations
I wanted to see the brain areas underlying the choice correlations in the
build-up and response time windows. I had hoped that in the build-up window the
posterior cingulate and other accumulated evidence areas would have the
strongest effects, while effects would concentrate on purely motor areas in the
response window.

Results are in:
	source_singledot_201808171756.h5

Unfortunately, the results are not as clear as I had wished. The areas found in
the build-up window indeed are consistent with the areas I found for
accumulated evidence, but the effects in right posterior cingulate and V6 are
much lower in the build-up window. Instead I see additional effects in
opercular areas (OP2-3) and more widespread effects in mid-cingulate motor
areas, premotor cortex and somatosensory and motor cortex. These effects are
even more prominent in the response window, as they should be. The effects also
more strongly stretch into anterior cingulate cortex. To my surprise there are
also reasonably strong effects in the whole of posterior cingulate cortex and
weaker effects in parietal areas.

Overall it looks as if the effects in the response window are just expanded
versions of those in the build-up window, although there is also clearly a
shift of the strongest effects from premotor areas to primary motor and
somatonsensory areas. I have a hard time consolidating the effects in posterior
cingulate cortex in the response window. Also, why do parietal areas (even if
only weakly) signal the choice at response time? This could mean two things:
Either parietal and posterior cingulate areas are generally involved in motor
preparation so that everything we see is some kind of "motor effect", or once
the decision and the corresponding response has been made, the perceptual areas
involved in determining the decision get feedback about the decision and signal
it together with the motor areas very strongly.