new analysis trying to replicate choice correlation differences in sources
The sensor topographies exhibited a relatively clear pattern of differences for
times before and around the response. I was interested in determining the
underlying differences in effects across brain areas and conducted the same
analysis in source space that I had done for the sensor topographies.

It turns out that it is much harder to get reliable results in source space
than in sensors. My current intuition is that this is because the effects
before the response are comparably weaker in source space than in sensor space,
i.e., the distribution of betas in brain areas is pretty close to 0 and noisy
before the response, but clearly above 0 around the response.

Because the signs of betas and effects changes across brain areas due to the
anatomy and because such sign reversals are also reflected in the differences,
I look at the magnitude differences. This makes the interpretation of the
resulting plots much easier as one colour always means that the effect was
stronger at the associated time point.

This analysis indicates that in all brain areas with sizable differences the
effects are stronger around the response than before the response, i.e., I
cannot really explain the differences in the topographies by activation of
different brain areas which is rather strange.

I tried a few things to investigate, or rather overcome this. Different
normalisations (scaling) so far did not change the results. Also I swapped
scaling and averaging of times within the time window, but this also did not
change much.

I have currently identified two more issues that could influence these
results: 1) The statistical analysis could be too noisy for the time points
before the response. 2) The normalisation does not work properly, because there
are too many brain areas with no effects that influence the mean magnitude. To
overcome 1) I think of running an expanded regression including all time points
from the time window. To overcome 2) I'm thinking of only using the brain areas
with significant effects for computing the scaling factor.