new analysis with aggregated data within response-aligned time windows
I wanted to make first_level beta estimates more reliable. So I shifted the
averaging into the regression by simply pooling data from time points that
fall into the selected time window. So in this analysis each time point enters
as an independent data point in the regression and regression coefficients
automatically represent the average coefficient within the time window.

I was able to implement all this in the infer_singledot infrastructure with
minimal changes.

Results of this run with response-aligned time windows build-up and response
can be found in:
	source_singledot_201808291410.h5