new function standardising normalisation of magnitudes
I thought more principally about how to normalise magnitudes of data and
decided that it is best to simultaneously normalise to min=0 and mean=1. The
issue with only normalising to mean magnitude=1 by scaling is that data sets
with magnitudes far from 0 will squish their variance considerably when they
get scaled to mean=1. So this normalisation will reduce differences between
magnitudes across the data set which is not what I want when comparing the
correlation patterns. To avoid this reduction in variance I first shift
magnitudes to min=0 before scaling to mean=1.

Note that the new function in helpers works on signed data, too.

For the differences between response-aligned time windows I did not notice any
differences for including the shift to min=0, or not.