major overhaul of the brain difference figure(s)
This is now all based on the windowed regression results and I first produce
figures showing the effects in the two time windows from this analysis. For the
difference figure itself I tried a few different ways of normalising the
magnitudes, but settled on the theoretically cleanest version as captured in
helpers.normalise_magnitudes. Additionally, I decided to compute the
normalisation only on brain areas with significant effects in the corresponding
time window, because otherwise the majority of brain areas with small effects
dominate the normalisation and make it harder to investigate the differences in
spatial activation patterns. Afterall I was mainly interested to see which
brain areas are activated in the one, but not the other time window.

I decreased fdr_alpha from 0.05 to 0.01, because I had a hard time interpreting
the (weak) significant activations for correlations that were larger in the
build-up window. These areas were: right - TPOJ1, MST, V7, 9p, 52, p24, 7PL,
EC; left - 7Pm, V3CD. So there are temporal, parietal, frontal and cingulate
areas which did not necessarily occur in the other analyses. From these, the
occipto-temporal and parietal activations make kind of sense, but what with the
others and why some in the left and some in the right hemisphere? So it is
likely that some of these are false positives, but which ones? Unfortunately
none survived fdr_alpha=0.01 to tell. So I rather stick with leaving the
question open in the paper, or rather only interpreting the stronger
differences favouring effects around the response time.

I stitched the three generated figures manually together using Gimp
(win_brain_response.xcf).