new figure comparing average signal for left and right responses
Hame pointed out that the central topography for the choice correlations
(response regressor) are rather unintuitive: Why should a right response lead
to increased signal over central sensors and a left response to increased
signal over lateral sensors? I wondered that myself before. So I decided to
investigate the issue further by directly checking the average signals
associated with left and right choices around the response time. These should
reflect some kind of lateralisation.

To use the infrastructure in place I formulated the analysis as regression and
created the two new left_response and right_response regressors which indicate
the corresponding trials. I originally wanted to run an analysis including
intercept, response and the two new regressors, but this is highly problematic.
Even though the individual left or right regressors could be used together with
response and intercept, their combination is collinear with both response and
intercept. So I had to exclude intercept and response regressors from the
analysis to make this work.

Inference results are in:
	meg_singledot_201809051639.h5

The average choice-specific signals display a topography presumably
corresponding to a single strong dipole over fronto-lateral sensors which
probably corresponds to the lateralised motor activation in primary motor and
somatosensory areas. Also the lateralisation of the dipoles speaks for that:
left choices have the dipole centre in the right hemisphere and the other way
around for the left hemisphere. The difference of right - left topographies
gives the central positivity, as expected.